+++ View/LoginWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="PinusPengger.ViewModel.LoginWindowController"
+﻿<Window x:Class="PinusPengger.View.LoginWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -19,9 +19,9 @@
             <RowDefinition Height="*"/>
             <RowDefinition Height="125"/>
         </Grid.RowDefinitions>
-        <Grid.ColumnDefinitions> 
+        <Grid.ColumnDefinitions>
             <ColumnDefinition Width="50"/>
-            <ColumnDefinition Width="25"/>  
+            <ColumnDefinition Width="25"/>
             <ColumnDefinition Width="*"/>
             <ColumnDefinition Width="75"/>
         </Grid.ColumnDefinitions>
@@ -180,6 +180,17 @@
                               Margin="0,0,0,1" />
                     </Grid>
                 </Button>
+                <Grid Grid.Row="5">
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition/>
+                        <ColumnDefinition Width="90"/>
+                    </Grid.ColumnDefinitions>
+                    <TextBlock Foreground="Red"
+                               Text="Invalid Password"
+                               Margin="10,0,0,0"
+                               Visibility="Hidden"
+                               />
+                </Grid>
             </Grid>
         </Border>
         <TextBlock FontSize="30"
