+++ View/MainWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="PinusPengger.ViewModel.MainWindowController"
+﻿<Window x:Class="PinusPengger.View.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -180,7 +180,7 @@
         <UserControl Grid.Row="2" Content="{Binding CurrentView}"
                      HorizontalAlignment="Center" VerticalAlignment="Top"
                      Height="420">
-            
+
         </UserControl>
     </Grid>
 </Window>
